Restructure code, add addAnswers button.
