+++ project/src/main/java/com/google/autograder/servlets/AddAnswersServlet.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Arrays;
 import com.google.autograder.data.Database;
+import com.google.autograder.data.Detect;
 import com.google.autograder.data.Answer;
 import com.google.gson.Gson;
 import java.io.IOException;
@@ -15,21 +16,31 @@ import com.google.appengine.api.datastore.Entity;
 import com.google.appengine.api.datastore.PreparedQuery;
 import com.google.appengine.api.datastore.Query;
 import com.google.appengine.api.datastore.Query.SortDirection;
+import java.io.File;
 
-public final class GroupServlet extends HttpServlet {
+public final class AddAnswersServlet extends HttpServlet {
 
   @Override
   public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-    // get all the answer entities based on answer + question keys
-    System.out.println("Creating groups...");
+    // add answer entities (hardcoded rn)
+    System.out.println("Adding answers...");
     String assignmentKey = request.getParameter("assignment-key");
     String questionKey = request.getParameter("question-key");
 
-    String json = Database.createGroups(assignmentKey, questionKey);
+    File folder = new File("images/answers");
+    File[] listOfFiles = folder.listFiles();
+
+    for (File file : listOfFiles) {
+        String path = "images/answers/" + file.getName();
+        Database.addAnswer(path, parseAnswer(path), 0, assignmentKey, questionKey);
+    }
     System.out.println("Done!");
-    System.out.println(json);
-    response.setContentType("application/json;");
-    response.getWriter().println(json);
+    return;
   }
-  
+
+  public String parseAnswer(String filePath) throws IOException {
+      Detect d = new Detect();
+      return d.detectDocumentText(filePath);
+
+  }
 }
