+++ project/src/main/webapp/pages/groups.html
@@ -11,8 +11,9 @@
 <body>
     <h1>Assignment Groups</h1>
     <h2>Group 1: 10/10</h2>
-    <div id="group-container"><script>getAnswers();</script></div>
+    <!-- <div id="group-container"><script>getAnswers();</script></div> -->
     <button onclick="computerGenerateGroups()">Computer Generate Groups</button>
+    <button onclick="addAnswers()">Add Answers</button>
 
 
 </body>
