+++ project/src/main/webapp/scripts/groups.js
@@ -13,6 +13,10 @@ function drop(ev) {
   ev.target.appendChild(document.getElementById(data));
 }
 
+async function addAnswers() {
+    var servletURL = "/addAnswers?assignment-key=" + getUrlVars()['assignment-key'] + "&question-key=" + getUrlVars()["question-key"];
+}
+
 async function getAnswers() {
     var servletURL = "/answer?assignment-key=" + getUrlVars()['assignment-key'] + "&question-key=" + getUrlVars()["question-key"];
